Remove old widget implementations
