+++ lib/main.dart
@@ -1,14 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:news_tracker/providers/tracked_term_provider_locked.dart';
 import 'package:news_tracker/screens/home_screen.dart';
 import 'package:news_tracker/utils/initialize_app.dart';
-import 'package:news_tracker/widgets/coreui/app_bar.dart';
-import 'package:news_tracker/widgets/coreui/drawer.dart';
-
-import 'widgets/page_body_container.dart';
-import 'widgets/tracked_terms/add_tracked_term.dart';
-import 'widgets/tracked_terms/tracked_terms_list.dart';
 
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
@@ -42,70 +35,7 @@ class NewsTracker extends StatelessWidget {
           surface: Color.fromARGB(237, 250, 240, 248),
         ),
       ),
-      home: HomeScreen(),
-      // MyHomePage(
-      //   showPermissionDialog: showPermissionDialog,
-      // ),
+      home: HomeScreen(showPermissionDialog: showPermissionDialog),
     );
-  }
-}
-
-class MyHomePage extends ConsumerWidget {
-  final bool showPermissionDialog;
-
-  const MyHomePage({required this.showPermissionDialog});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final termsAsync = ref.watch(newTrackedTermsProvider);
-
-    final termsCount = termsAsync.when(
-      data: (terms) => terms.length,
-      loading: () => 0,
-      error: (_, _) => 0,
-    );
-
-    if (showPermissionDialog) {
-      permissionCallback(context);
-    }
-
-    return Scaffold(
-      appBar: DefaultBar(),
-      drawer: OptionsDrawer(),
-      body: PageBodyContainer(
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: [
-          Text('Term count: $termsCount'),
-          Expanded(child: TrackedTermsList()),
-          Padding(
-            padding: const EdgeInsets.only(bottom: 32.0),
-            child: AddTrackedTerm(),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// Inform user to enable notifications in settings if permanently disabled
-  void permissionCallback(BuildContext context) {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      showDialog(
-        context: context,
-        builder: (context) => AlertDialog(
-          title: Text('Notification Permission'),
-          content: Text(
-            'Notification permission is permanently denied. NewsTracker will not '
-            'work properly without notification permission. Visit your phone\'s '
-            'Settings -> Apps -> NewsTracker -> Permissions to enable',
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(),
-              child: Text('OK'),
-            ),
-          ],
-        ),
-      );
-    });
   }
 }
