+++ lib/new_widgets/home_screen/term_inputs_widget.dart
