+++ lib/new_widgets/home_screen/term_tile_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../model/tracked_term.dart';
+import '../../model/tracked_term.dart';
 
 class TermTile extends StatelessWidget {
   final TrackedTerm term;
